pyproject.toml
[project]
name = "platform-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.5.0",
    "aiokafka>=0.13.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-instrumentation-asyncpg>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.40.0",
]
